move performance integration tests to a separate file and update make integration-test target to run this test independently
