+++ tests/integration/docker-compose-performance.yml
@@ -6,37 +6,37 @@ services:
     container_name: "mysql_perf_8-0-40"
     restart: always
     environment:
-      MYSQL_ROOT_PASSWORD: DBpwd1234
+      MYSQL_ALLOW_EMPTY_PASSWORD: true
     build:
       context: ../../
       dockerfile: tests/integration/mysql-performance-config/versions/8.0.40/Dockerfile
     links:
-      - nri-mysql
+      - nri-mysql-perf
 
   mysql_perf_8-4-0:
     container_name: "mysql_perf_8-4-0"
     restart: always
     environment:
-      MYSQL_ROOT_PASSWORD: DBpwd1234
+      MYSQL_ALLOW_EMPTY_PASSWORD: true
     build:
       context: ../../
       dockerfile: tests/integration/mysql-performance-config/versions/8.4.0/Dockerfile
     links:
-      - nri-mysql
+      - nri-mysql-perf
 
   mysql_perf_latest-supported:
     container_name: "mysql_perf_latest-supported"
     restart: always
     environment:
-      MYSQL_ROOT_PASSWORD: DBpwd1234
+      MYSQL_ALLOW_EMPTY_PASSWORD: true
     build:
       context: ../../
       dockerfile: tests/integration/mysql-performance-config/versions/9.1.0/Dockerfile
     links:
-      - nri-mysql
+      - nri-mysql-perf
 
-  nri-mysql:
-    container_name: integration_nri-mysql_1
+  nri-mysql-perf:
+    container_name: integration_nri-mysql_perf_1
     build:
       context: ../../
       dockerfile: tests/integration/Dockerfile
